Make frame number uint32_t in PHY and MAC classes, update README and various fixes (#98)
* Make frame number uint32_t in PHY and MAC classes

* Fix logs

+++ src/mmwave/model/mmwave-ue-mac.cc
@@ -637,7 +637,7 @@ std::map<uint32_t, struct MacPduInfo>::iterator MmWaveUeMac::AddToMacPduMap (Dci
   uint8_t slotNum = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) % m_phyMacConfig->GetSlotsPerSubframe ();
   uint8_t deltaSubframe = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) / m_phyMacConfig->GetSlotsPerSubframe ();
   uint8_t sfNum = (m_sfNum + deltaSubframe) % m_phyMacConfig->GetSubframesPerFrame ();
-  uint16_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
+  uint32_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
 
   NS_ASSERT ((slotNum < m_phyMacConfig->GetSlotsPerSubframe ()) && (sfNum < m_phyMacConfig->GetSubframesPerFrame ())
              && (deltaSubframe >= 0) && (slotNum >= 0) && (sfNum >= 0) && (frameNum >= m_frameNum));
@@ -704,7 +704,7 @@ MmWaveUeMac::DoReceiveControlMessage  (Ptr<MmWaveControlMessage> msg)
                     uint8_t slotNum = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) % m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t deltaSubframe = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) / m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t sfNum = (m_sfNum + deltaSubframe) % m_phyMacConfig->GetSubframesPerFrame ();
-                    uint16_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
+                    uint32_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
 
                     NS_ASSERT ((slotNum < m_phyMacConfig->GetSlotsPerSubframe ()) && (sfNum < m_phyMacConfig->GetSubframesPerFrame ())
                                 && (deltaSubframe >= 0) && (slotNum >= 0) && (sfNum >= 0) && (frameNum >= m_frameNum));
@@ -901,7 +901,7 @@ MmWaveUeMac::DoReceiveControlMessage  (Ptr<MmWaveControlMessage> msg)
                     uint8_t slotNum = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) % m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t deltaSubframe = (m_slotNum + m_phyMacConfig->GetUlSchedDelay ()) / m_phyMacConfig->GetSlotsPerSubframe ();
                     uint8_t sfNum = (m_sfNum + deltaSubframe) % m_phyMacConfig->GetSubframesPerFrame ();
-                    uint16_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
+                    uint32_t frameNum = m_frameNum + ((m_sfNum + deltaSubframe) / m_phyMacConfig->GetSubframesPerFrame ());
 
                     NS_ASSERT ((slotNum < m_phyMacConfig->GetSlotsPerSubframe ()) && (sfNum < m_phyMacConfig->GetSubframesPerFrame ())
                                 && (deltaSubframe >= 0) && (slotNum >= 0) && (sfNum >= 0) && (frameNum >= m_frameNum));
